Split Instances into several files
