+++ lisa-sets/src/main/scala/lisa/maths/settheory/functions/DirectPreimages.scala
@@ -1,25 +1,47 @@
-package lisa.maths.topology
-
-import lisa.automation.settheory.SetTheoryTactics.*
-import lisa.maths.Quantifiers.*
+package lisa.maths.settheory.functions
 
 import lisa.automation.kernel.CommonTactics.Definition
-
-import lisa.maths.topology.Topology.*
-import lisa.maths.settheory.*
-import lisa.maths.settheory.SetTheory.*
+import lisa.automation.settheory.SetTheoryTactics.*
+import lisa.maths.Quantifiers.*
 import lisa.maths.settheory.SetTheoryBasics.*
 import lisa.automation.kernel.CommonTactics.*
 import lisa.maths.settheory.functions.Functionals.*
 import lisa.automation.settheory.SetTheoryTactics.UniqueComprehension
 import lisa.automation.settheory.SetTheoryTactics.TheConditional
+import lisa.maths.settheory.SetTheory.*
 
-object Instances extends lisa.Main {
-  import lisa.maths.settheory.SetTheory.*
+object DirectPreimages extends lisa.Main {
+
   // var defs
   private val x, y, z, a, b, c, t, p, f, r, s = variable
   private val X, T, T1, T2 = variable
   private val S, A, B, Y, o, O, O2, O3 = variable
+
+  /**
+   * Don't know why, but I need to paste it directly here otherwise we have a matcherror null error (from FunctionProperties)
+   */
+  val surjective = DEF(f, x, y) --> functionFrom(f, x, y) /\ ∀(b, in(b, y) ==> ∃(a, in(pair(a, b), f)))
+
+  /* Also copied the theorem (from FunctionProperties) */
+  val surjectiveImpliesRangeIsCodomain = Theorem(
+    surjective(f, x, y) |- (y === functionRange(f))
+  ) {
+    have(surjective(f, x, y) |- ∀(b, in(b, y) ==> ∃(a, in(pair(a, b), f)))) by Tautology.from(surjective.definition)
+    val surjDef = thenHave(surjective(f, x, y) |- in(b, y) ==> ∃(a, in(pair(a, b), f))) by InstantiateForall(b)
+    have(∀(t, in(t, functionRange(f)) <=> (∃(a, in(pair(a, t), f))))) by InstantiateForall(functionRange(f))(functionRange.definition of (r -> f))
+    val rangeDef = thenHave(in(b, functionRange(f)) <=> (∃(a, in(pair(a, b), f)))) by InstantiateForall(b)
+
+    have(surjective(f, x, y) |- in(b, y) ==> in(b, functionRange(f))) by Tautology.from(surjDef, rangeDef)
+    thenHave(surjective(f, x, y) |- ∀(b, in(b, y) ==> in(b, functionRange(f)))) by RightForall
+    val surjsub = andThen(Substitution.applySubst(subsetAxiom of (x -> y, y -> functionRange(f))))
+
+    have((surjective(f, x, y), functionFrom(f, x, y)) |- subset(y, functionRange(f)) /\ subset(functionRange(f), y)) by RightAnd(surjsub, functionImpliesRangeSubsetOfCodomain)
+    val funceq = andThen(Substitution.applySubst(subsetEqualitySymmetry of (x -> y, y -> functionRange(f))))
+
+    val surjfunc = have(surjective(f, x, y) |- functionFrom(f, x, y)) by Tautology.from(surjective.definition)
+
+    have(thesis) by Cut(surjfunc, funceq)
+  }
 
   inline def directImageFormula = y ∈ s <=> (y ∈ Y /\ ∃(x, (app(f, x) === y) /\ x ∈ A))
 
@@ -132,23 +154,6 @@ object Instances extends lisa.Main {
     thenHave(∀(z, z ∈ directImage(f, X, Y, setUnion(A, B)) <=> z ∈ setUnion(directImage(f, X, Y, A), directImage(f, X, Y, B)))) by RightForall
     andThen(Substitution.applySubst(extensionalityAxiom of (x := directImage(f, X, Y, setUnion(A, B)), y := setUnion(directImage(f, X, Y, A), directImage(f, X, Y, B)))))
   }
-
-  /*inline def directImageUnionFormula = y ∈ s <=> y ∈ Y /\ ∃(A, y ∈ directImage(f, A))
-  val directImageUnionUniqueness = Theorem(
-    functional(f) |- ∃!(s, ∀(y, directImageUnionFormula))
-  ) {
-    have(∃!(s, ∀(y, directImageUnionFormula))) by UniqueComprehension(Y, lambda(y, ∃(A, y ∈ directImage(f, A))))
-    thenHave(thesis) by Weakening
-  }
-
-  val directImageUnion = DEF(f, A) --> TheConditional(s, forall(z, z ∈ s <=> ∃(A, z ∈ directImage(f, A))))(directImageUnionUniqueness)
-
-  val directImageUnionThm = Theorem(
-    functional(f) /\ forall(A, A ∈ T ==> subset(A, X)) |-
-      union(directImage(f, T)) === directImage(f, union(A))
-  ) {
-    sorry
-  }*/
 
   inline def preimageFormula = x ∈ s <=> (x ∈ X /\ app(f, x) ∈ B)
 
@@ -339,32 +344,6 @@ object Instances extends lisa.Main {
       preimage(f, X, Y, union(B)) === union(preimages(f, X, Y, B))
   ) {
     sorry
-  }
-
-  // Couldn't import surjectivity from FunctionProperties without an error, so here it is
-  val surjective = DEF(f, x, y) --> functionFrom(f, x, y) /\ ∀(b, in(b, y) ==> ∃(a, in(pair(a, b), f)))
-
-  /**
-   * Theorem --- if a function is [[surjective]], its range is equal to its codomain.
-   */
-  val surjectiveImpliesRangeIsCodomain = Theorem(
-    surjective(f, x, y) |- (y === functionRange(f))
-  ) {
-    have(surjective(f, x, y) |- ∀(b, in(b, y) ==> ∃(a, in(pair(a, b), f)))) by Tautology.from(surjective.definition)
-    val surjDef = thenHave(surjective(f, x, y) |- in(b, y) ==> ∃(a, in(pair(a, b), f))) by InstantiateForall(b)
-    have(∀(t, in(t, functionRange(f)) <=> (∃(a, in(pair(a, t), f))))) by InstantiateForall(functionRange(f))(functionRange.definition of (r -> f))
-    val rangeDef = thenHave(in(b, functionRange(f)) <=> (∃(a, in(pair(a, b), f)))) by InstantiateForall(b)
-
-    have(surjective(f, x, y) |- in(b, y) ==> in(b, functionRange(f))) by Tautology.from(surjDef, rangeDef)
-    thenHave(surjective(f, x, y) |- ∀(b, in(b, y) ==> in(b, functionRange(f)))) by RightForall
-    val surjsub = andThen(Substitution.applySubst(subsetAxiom of (x -> y, y -> functionRange(f))))
-
-    have((surjective(f, x, y), functionFrom(f, x, y)) |- subset(y, functionRange(f)) /\ subset(functionRange(f), y)) by RightAnd(surjsub, functionImpliesRangeSubsetOfCodomain)
-    val funceq = andThen(Substitution.applySubst(subsetEqualitySymmetry of (x -> y, y -> functionRange(f))))
-
-    val surjfunc = have(surjective(f, x, y) |- functionFrom(f, x, y)) by Tautology.from(surjective.definition)
-
-    have(thesis) by Cut(surjfunc, funceq)
   }
 
   val directImageEmptySet = Theorem(
@@ -561,439 +540,5 @@ object Instances extends lisa.Main {
       directImageX,
       equalityTransitivity of (x := Y, y := functionRange(f), z := directImage(f, X, Y, X))
     )
-  }
-
-  // -------------------
-  // Mappings
-  // -------------------
-  val mapping = DEF(f, X, T1, Y, T2) -->
-    (functionFrom(f, X, Y) /\ topology(X, T1) /\ topology(Y, T2))
-
-  // -------------------
-  // Continuity
-  // -------------------
-  val continuous = DEF(f, X, T1, Y, T2) -->
-    (mapping(f, X, T1, Y, T2) /\ forall(O, O ∈ T2 ==> preimage(f, X, Y, O) ∈ T1))
-
-  // -------------------
-  // Connectedness
-  // -------------------
-  val clopen = DEF(X, T, A) --> (
-    topology(X, T) /\
-      A ∈ T /\ setDifference(X, A) ∈ T
-  )
-
-  val connectedTop = DEF(X, T) --> (
-    topology(X, T) /\
-      forall(A, clopen(X, T, A) ==> ((A === emptySet) \/ (A === X)))
-  )
-
-  // -------------------
-  // Intermediate value theorem
-  // -------------------
-  val intermediateValueThm = Theorem((connectedTop(X, T1), continuous(f, X, T1, Y, T2), surjective(f, X, Y)) |- connectedTop(Y, T2)) {
-    assume(connectedTop(X, T1), continuous(f, X, T1, Y, T2), surjective(f, X, Y))
-    val xIsTop = have(topology(X, T1)) by Tautology.from(continuous.definition, mapping.definition)
-    val yIsTop = have(topology(Y, T2)) by Tautology.from(continuous.definition, mapping.definition)
-
-    val xIsConnected = have(forall(A, clopen(X, T1, A) ==> ((A === emptySet) \/ (A === X)))) by Tautology.from(connectedTop.definition of (T := T1))
-    val isContinuous = have(forall(O, O ∈ T2 ==> preimage(f, X, Y, O) ∈ T1)) by Tautology.from(continuous.definition)
-
-    val fIsFunction = have(functionFrom(f, X, Y)) by Tautology.from(continuous.definition, mapping.definition)
-
-    have(clopen(Y, T2, A) ==> ((A === emptySet) \/ (A === Y))) subproof {
-      assume(clopen(Y, T2, A))
-
-      val aIsSubset = have(A ⊆ Y) subproof {
-        have(A ∈ T2) by Tautology.from(clopen.definition of (X := Y, T := T2))
-        have(A ∈ powerSet(Y)) by Tautology.from(
-          lastStep,
-          yIsTop,
-          topology.definition of (X := Y, T := T2),
-          setOfSubsets.definition of (X := Y, T := T2),
-          subsetTactic of (x := T2, y := powerSet(Y), z := A)
-        )
-        have(thesis) by Tautology.from(lastStep, powerAxiom of (x := A, y := Y))
-      }
-
-      val preimageA = have(A ∈ T2 ==> preimage(f, X, Y, A) ∈ T1) by InstantiateForall(A)(isContinuous)
-      val yIsClopen = have(A ∈ T2 /\ setDifference(Y, A) ∈ T2) by Tautology.from(clopen.definition of (X := Y, T := T2))
-      val part1 = have(preimage(f, X, Y, A) ∈ T1) by Tautology.from(yIsClopen, preimageA)
-      val preimageYA = have(setDifference(Y, A) ∈ T2 ==> preimage(f, X, Y, setDifference(Y, A)) ∈ T1) by InstantiateForall(setDifference(Y, A))(isContinuous)
-      have(preimage(f, X, Y, setDifference(Y, A)) ∈ T1) by Tautology.from(yIsClopen, preimageYA)
-      val part2 = have(setDifference(X, preimage(f, X, Y, A)) ∈ T1) by Tautology.from(
-        lastStep,
-        aIsSubset,
-        fIsFunction,
-        preimageDifference,
-        replaceEqualityContainsLeft of (x := setDifference(X, preimage(f, X, Y, A)), y := preimage(f, X, Y, setDifference(Y, A)), z := T1)
-      )
-
-      // So f^-1(A) is clopen
-      val inverseIsClopen = have(clopen(X, T1, preimage(f, X, Y, A))) by Tautology.from(
-        xIsTop,
-        part1,
-        part2,
-        clopen.definition of (T := T1, A := preimage(f, X, Y, A))
-      )
-
-      // Hence (f^-1(A) === emptySet) \/ (preimage(f, X, Y, A) === X) by connectedness of X
-      have(clopen(X, T1, preimage(f, X, Y, A)) ==> (preimage(f, X, Y, A) === emptySet) \/ (preimage(f, X, Y, A) === X)) by InstantiateForall(preimage(f, X, Y, A))(xIsConnected)
-      val preImageIsConnected = have((preimage(f, X, Y, A) === emptySet) \/ (preimage(f, X, Y, A) === X)) by Tautology.from(
-        lastStep,
-        inverseIsClopen
-      )
-
-      // Use the fact that f(emptyset)=emptyset, f(f^-1(A)) = A (by surjectivity), f(X) = Y (by surjectivity) to conclude
-      val firstCase = have(preimage(f, X, Y, A) === emptySet |- A === emptySet) subproof {
-        assume(preimage(f, X, Y, A) === emptySet)
-        have(thesis) by Tautology.from(
-          lastStep,
-          fIsFunction,
-          aIsSubset,
-          applyDirectImage of (A := preimage(f, X, Y, A), B := emptySet),
-          directImagePreimageSurjective,
-          directImageEmptySet,
-          equalityTransitivity of (x := directImage(f, X, Y, preimage(f, X, Y, A)), y := directImage(f, X, Y, emptySet), z := emptySet),
-          equalitySymmetry of (x := directImage(f, X, Y, preimage(f, X, Y, A)), y := A),
-          equalityTransitivity of (x := A, y := directImage(f, X, Y, preimage(f, X, Y, A)), z := emptySet)
-        )
-      }
-
-      val secondCase = have(preimage(f, X, Y, A) === X |- A === Y) subproof {
-        assume(preimage(f, X, Y, A) === X)
-        have(thesis) by Tautology.from(
-          lastStep,
-          fIsFunction,
-          aIsSubset,
-          applyDirectImage of (A := preimage(f, X, Y, A), B := X),
-          directImagePreimageSurjective,
-          directImageX,
-          surjectiveImpliesRangeIsCodomain of (x := X, y := Y),
-          equalitySymmetry of (x := Y, y := functionRange(f)),
-          equalityTransitivity of (x := directImage(f, X, Y, X), y := functionRange(f), z := Y),
-          equalityTransitivity of (x := directImage(f, X, Y, preimage(f, X, Y, A)), y := directImage(f, X, Y, X), z := Y),
-          equalitySymmetry of (x := directImage(f, X, Y, preimage(f, X, Y, A)), y := A),
-          equalityTransitivity of (x := A, y := directImage(f, X, Y, preimage(f, X, Y, A)), z := Y)
-        )
-      }
-
-      have(thesis) by Tautology.from(preImageIsConnected, firstCase, secondCase)
-    }
-
-    val allClopen = thenHave(forall(A, clopen(Y, T2, A) ==> ((A === emptySet) \/ (A === Y)))) by RightForall
-    have(connectedTop(Y, T2)) by Tautology.from(allClopen, yIsTop, connectedTop.definition of (X := Y, T := T2))
-  }
-
-  // -------------------
-  // Compactness
-  // -------------------
-
-  val cover = DEF(X, O) -->
-    forall(o, in(o, O) ==> subset(o, X)) /\
-    subset(X, union(O))
-
-  val openCover = DEF(X, T, O) -->
-    cover(X, O) /\ subset(O, T)
-
-  val finite = DEF(X) --> (X === emptySet) // TODO
-
-  val compact = DEF(X, T) -->
-    topology(X, T) /\
-    forall(
-      O,
-      openCover(X, T, O) ==>
-        ∃(
-          O2, // Another subcovering
-          subset(O2, O) /\ cover(X, O2) /\ finite(O2)
-        )
-    )
-
-  val coverDirectImage = Theorem(
-    (functionFrom(f, X, Y), A ⊆ powerSet(X), cover(X, A)) |- cover(directImage(f, X, Y, X), directImages(f, X, Y, A))
-  ) {
-    assume(functionFrom(f, X, Y), A ⊆ powerSet(X), cover(X, A))
-
-    sorry
-  }
-
-  /* Any subset of an open cover is an open cover */
-  val subsetOpenCover = Theorem(
-    (openCover(X, T, O), subset(O2, O), cover(X, O2)) |- openCover(X, T, O2)
-  ) {
-    assume(openCover(X, T, O), O2 ⊆ O, cover(X, O2))
-
-    have(thesis) by Tautology.from(
-      openCover.definition of (O := O2),
-      openCover.definition,
-      subsetTransitivity of (a := O2, b := O, c := T)
-    )
-  }
-
-  /* The preimages of some set in P(Y) are in P(X) */
-  val preimagesInPowerSet = Theorem(
-    (functionFrom(f, X, Y), A ⊆ powerSet(Y)) |- preimages(f, X, Y, A) ⊆ powerSet(X)
-  ) {
-    assume(functionFrom(f, X, Y), A ⊆ powerSet(Y))
-
-    have(x ∈ preimages(f, X, Y, A) ==> x ∈ powerSet(X)) by Tautology.from(
-      preimagesMembership of (A := A, x := x)
-    )
-    thenHave(forall(x, x ∈ preimages(f, X, Y, A) ==> x ∈ powerSet(X))) by RightForall
-    have(thesis) by Tautology.from(lastStep, subsetAxiom of (x := preimages(f, X, Y, A), y := powerSet(X)))
-  }
-
-  /* The set of direct images is finite */
-  val directImageFinite = Theorem(
-    (functionFrom(f, X, Y), A ⊆ powerSet(X), finite(A)) |- finite(directImages(f, X, Y, A))
-  ) {
-    assume(functionFrom(f, X, Y), A ⊆ powerSet(X), finite(A))
-    // TODO: Needs to have a notion of finiteness to complete the proof
-    // Normally it should just be because there is a bijection between directImages(f, X, Y, A) and A, and A is finite
-    sorry
-  }
-
-  val heineBorelThm = Theorem((compact(X, T1), continuous(f, X, T1, Y, T2), surjective(f, X, Y)) |- compact(Y, T2)) {
-    assume(compact(X, T1), continuous(f, X, T1, Y, T2), surjective(f, X, Y))
-
-    val xIsTop = have(topology(X, T1)) by Tautology.from(continuous.definition, mapping.definition)
-    val yIsTop = have(topology(Y, T2)) by Tautology.from(continuous.definition, mapping.definition)
-
-    val xIsCompact = have(forall(O, openCover(X, T1, O) ==> ∃(O2, subset(O2, O) /\ cover(X, O2) /\ finite(O2)))) by Tautology.from(
-      compact.definition of (T := T1)
-    )
-    val isContinuous = have(forall(O, O ∈ T2 ==> preimage(f, X, Y, O) ∈ T1)) by Tautology.from(continuous.definition)
-
-    val fIsFunction = have(functionFrom(f, X, Y)) by Tautology.from(continuous.definition, mapping.definition)
-
-    have(openCover(Y, T2, O) |- ∃(O2, subset(O2, O) /\ cover(Y, O2) /\ finite(O2))) subproof {
-      assume(openCover(Y, T2, O))
-
-      have(forall(O, (O ⊆ T2) ==> (union(O) ∈ T2))) by Tautology.from(
-        containsUnion.definition of (T := T2),
-        yIsTop,
-        topology.definition of (X := Y, T := T2)
-      )
-      thenHave(O ⊆ T2 ==> (union(O) ∈ T2)) by InstantiateForall(O)
-      val unionInT2 = have(union(O) ∈ T2) by Tautology.from(
-        openCover.definition of (X := Y, T := T2),
-        lastStep
-      )
-
-      val oInPowerSet = have(O ⊆ powerSet(Y)) by Tautology.from(
-        openCover.definition of (X := Y, T := T2),
-        yIsTop,
-        topology.definition of (X := Y, T := T2),
-        setOfSubsets.definition of (X := Y, T := T2),
-        subsetTransitivity of (a := O, b := T2, c := powerSet(Y))
-      )
-      val unionOSubsetY = have(union(O) ⊆ Y) by Tautology.from(
-        oInPowerSet,
-        subsetClosedUnion of (x := O, y := Y)
-      )
-
-      // We have an open cover of X, that's preimages(f, X, Y, O)
-      val isOpenCover = have(openCover(X, T1, preimages(f, X, Y, O))) subproof {
-        // Firstly, it's a cover
-        val isCover = have(cover(X, preimages(f, X, Y, O))) subproof {
-          have(o ∈ preimages(f, X, Y, O) ==> o ⊆ X) subproof {
-            assume(o ∈ preimages(f, X, Y, O))
-            have(o ∈ powerSet(X)) by Tautology.from(
-              preimagesMembership of (A := O, x := o),
-              fIsFunction,
-              oInPowerSet
-            )
-            have(o ⊆ X) by Tautology.from(lastStep, powerAxiom of (x := o, y := X))
-          }
-          val firstPart = thenHave(∀(o, o ∈ preimages(f, X, Y, O) ==> o ⊆ X)) by RightForall
-
-          // The covering part
-          have(x ∈ X ==> x ∈ union(preimages(f, X, Y, O))) subproof {
-            assume(x ∈ X)
-            // Function application
-            have(app(f, x) ∈ Y) by Tautology.from(
-              fIsFunction,
-              lastStep,
-              functionFromApplication of (x := X, y := Y, a := x),
-              functionFrom.definition of (x := X, y := Y)
-            )
-            // Since Y is covered by O
-            have(app(f, x) ∈ union(O)) by Tautology.from(
-              lastStep,
-              openCover.definition of (X := Y, T := T2),
-              cover.definition of (X := Y),
-              subsetTactic of (x := Y, y := union(O), z := app(f, x))
-            )
-            // That's the definition of being in the preimage!
-            have(x ∈ preimage(f, X, Y, union(O))) by Tautology.from(
-              lastStep,
-              preimageMembership of (B := union(O)),
-              fIsFunction,
-              unionOSubsetY
-            )
-            // use that preimage(f, X, Y, union(O)) ⊆ union(preimages(f, X, Y, O))
-            have(x ∈ union(preimages(f, X, Y, O))) by Tautology.from(
-              lastStep,
-              preimageUnionThm of (B := O),
-              replaceEqualityContainsRight of (x := preimage(f, X, Y, union(O)), y := union(preimages(f, X, Y, O)), z := x),
-              fIsFunction,
-              oInPowerSet
-            )
-          }
-          thenHave(forall(x, x ∈ X ==> x ∈ union(preimages(f, X, Y, O)))) by RightForall
-          val secondPart = have(subset(X, union(preimages(f, X, Y, O)))) by Tautology.from(
-            lastStep,
-            subsetAxiom of (x := X, y := union(preimages(f, X, Y, O)))
-          )
-
-          have(thesis) by Tautology.from(firstPart, secondPart, cover.definition of (O := preimages(f, X, Y, O)))
-        }
-
-        // Also, its elements are open
-        have(z ∈ preimages(f, X, Y, O) ==> z ∈ T1) subproof {
-          assume(z ∈ preimages(f, X, Y, O))
-          val existsa = have(∃(a, a ∈ O /\ (z === preimage(f, X, Y, a)))) by Tautology.from(
-            lastStep,
-            preimagesMembership of (A := O, x := z),
-            fIsFunction,
-            oInPowerSet
-          )
-          have(a ∈ O /\ (z === preimage(f, X, Y, a)) |- z ∈ T1) subproof {
-            assume(a ∈ O /\ (z === preimage(f, X, Y, a)))
-            val aInT2 = have(a ∈ T2) by Tautology.from(
-              openCover.definition of (X := Y, T := T2),
-              subsetTactic of (x := O, y := T2, z := a)
-            )
-            have(a ∈ T2 ==> preimage(f, X, Y, a) ∈ T1) by InstantiateForall(a)(isContinuous)
-            have(preimage(f, X, Y, a) ∈ T1) by Tautology.from(
-              aInT2,
-              lastStep
-            )
-            have(z ∈ T1) by Tautology.from(lastStep, replaceEqualityContainsLeft of (x := z, y := preimage(f, X, Y, a), z := T1))
-          }
-          thenHave(∃(a, a ∈ O /\ (z === preimage(f, X, Y, a))) |- z ∈ T1) by LeftExists
-          have(thesis) by Tautology.from(existsa, lastStep)
-        }
-        thenHave(forall(z, z ∈ preimages(f, X, Y, O) ==> z ∈ T1)) by RightForall
-        val isOpenSubset = have(preimages(f, X, Y, O) ⊆ T1) by Tautology.from(
-          subsetAxiom of (x := preimages(f, X, Y, O), y := T1),
-          lastStep
-        )
-
-        have(thesis) by Tautology.from(openCover.definition of (T := T1, O := preimages(f, X, Y, O)), isCover, isOpenSubset)
-      }
-
-      // Whence the existence of a finite subcover O3
-      have(openCover(X, T1, preimages(f, X, Y, O)) ==> ∃(O3, subset(O3, preimages(f, X, Y, O)) /\ cover(X, O3) /\ finite(O3))) by InstantiateForall(preimages(f, X, Y, O))(
-        xIsCompact
-      )
-      val existsO3 = have(∃(O3, subset(O3, preimages(f, X, Y, O)) /\ cover(X, O3) /\ finite(O3))) by Tautology.from(lastStep, isOpenCover)
-
-      have(
-        O3 ⊆ preimages(f, X, Y, O) /\ cover(X, O3) /\ finite(O3)
-          |-
-          subset(directImages(f, X, Y, O3), O) /\ cover(Y, directImages(f, X, Y, O3)) /\ finite(directImages(f, X, Y, O3))
-      ) subproof {
-        assume(O3 ⊆ preimages(f, X, Y, O), cover(X, O3), finite(O3))
-
-        val o3InPowerSet = have(O3 ⊆ powerSet(X)) subproof {
-          have(preimages(f, X, Y, O) ⊆ powerSet(X)) by Tautology.from(
-            fIsFunction,
-            oInPowerSet,
-            preimagesInPowerSet of (A := O)
-          )
-          have(thesis) by Tautology.from(lastStep, subsetTransitivity of (a := O3, b := preimages(f, X, Y, O), c := powerSet(X)))
-        }
-
-        // So it's a subset
-        have(z ∈ directImages(f, X, Y, O3) ==> z ∈ O) subproof {
-          assume(z ∈ directImages(f, X, Y, O3))
-
-          have(z ∈ powerSet(Y) /\ ∃(a, a ∈ O3 /\ (z === directImage(f, X, Y, a)))) by Tautology.from(
-            directImagesMembership of (A := O3, y := z),
-            fIsFunction,
-            o3InPowerSet
-          )
-
-          have(a ∈ O3 /\ (z === directImage(f, X, Y, a)) |- z ∈ O) subproof {
-            assume(a ∈ O3, z === directImage(f, X, Y, a))
-            val aInPreimages = have(a ∈ preimages(f, X, Y, O)) by Tautology.from(
-              lastStep,
-              subsetTactic of (x := O3, y := preimages(f, X, Y, O), z := a)
-            )
-            have(b ∈ O /\ (a === preimage(f, X, Y, b)) |- directImage(f, X, Y, a) ∈ O) subproof {
-              assume(b ∈ O, a === preimage(f, X, Y, b))
-              have(b ⊆ Y) by Tautology.from(
-                oInPowerSet,
-                subsetTactic of (z := b, x := O, y := powerSet(Y)),
-                powerAxiom of (x := b, y := Y)
-              )
-              val statement = have(directImage(f, X, Y, preimage(f, X, Y, b)) === b) by Tautology.from(
-                lastStep,
-                directImagePreimageSurjective of (A := b),
-                fIsFunction
-              )
-              thenHave(directImage(f, X, Y, a) === b) by RightSubstEq.withParametersSimple(List((a, preimage(f, X, Y, b))), lambda(x, directImage(f, X, Y, x) === b))
-              have(thesis) by Tautology.from(lastStep, replaceEqualityContainsLeft of (x := directImage(f, X, Y, a), y := b, z := O))
-            }
-            thenHave(exists(b, b ∈ O /\ (a === preimage(f, X, Y, b))) |- directImage(f, X, Y, a) ∈ O) by LeftExists
-            // use that (functionFrom(f, X, Y), O ⊆ powerSet(Y)) |- a ∈ preimages(f, X, Y, O) ==> (∃(b, b ∈ O /\ (a === preimage(f, X, Y, b)))))
-            have(directImage(f, X, Y, a) ∈ O) by Tautology.from(lastStep, aInPreimages, preimagesMembership of (A := O, x := a), fIsFunction, oInPowerSet)
-            // Conclude since z === directImage(f, X, Y, a)
-            have(z ∈ O) by Tautology.from(lastStep, replaceEqualityContainsLeft of (x := z, y := directImage(f, X, Y, a), z := O))
-          }
-
-          thenHave(∃(a, a ∈ O3 /\ (z === directImage(f, X, Y, a))) |- z ∈ O) by LeftExists
-
-          // Since z is in directImages, then we precisely have ∃(a, a ∈ O3 /\ (z === directImage(f, X, Y, a))) by `directImagesMembership`
-          have(thesis) by Tautology.from(lastStep, directImagesMembership of (A := O3, y := z), fIsFunction, o3InPowerSet)
-        }
-        thenHave(forall(z, z ∈ directImages(f, X, Y, O3) ==> z ∈ O)) by RightForall
-        val isSubset = have(directImages(f, X, Y, O3) ⊆ O) by Tautology.from(lastStep, subsetAxiom of (x := directImages(f, X, Y, O3), y := O))
-
-        // That is also covering Y
-        // We use that f is surjective to get that directImage(f, X, Y, X) = Y
-        val replacement = have(directImage(f, X, Y, X) === Y) by Tautology.from(imageSurjective, fIsFunction)
-        val coveringStatement = have(cover(directImage(f, X, Y, X), directImages(f, X, Y, O3))) by Tautology.from(
-          coverDirectImage of (A := O3),
-          fIsFunction,
-          o3InPowerSet
-        )
-        have(
-          ((directImage(f, X, Y, X) === Y), cover(directImage(f, X, Y, X), directImages(f, X, Y, O3)))
-            |- cover(Y, directImages(f, X, Y, O3))
-        ) subproof {
-          have(
-            ((directImage(f, X, Y, X) === Y), cover(directImage(f, X, Y, X), directImages(f, X, Y, O3)))
-              |- cover(directImage(f, X, Y, X), directImages(f, X, Y, O3))
-          ) by Tautology
-          thenHave(thesis) by RightSubstEq.withParametersSimple(List((directImage(f, X, Y, X), Y)), lambda(x, cover(x, directImages(f, X, Y, O3))))
-        }
-        val covering = have(cover(Y, directImages(f, X, Y, O3))) by Tautology.from(lastStep, coveringStatement, replacement)
-
-        // Finally it's finite since O3 is
-        val finiteCover = have(finite(directImages(f, X, Y, O3))) by Tautology.from(
-          directImageFinite of (A := O3),
-          fIsFunction,
-          o3InPowerSet
-        )
-
-        have(thesis) by Tautology.from(isSubset, finiteCover, covering)
-      }
-
-      thenHave(
-        subset(O3, preimages(f, X, Y, O)) /\ cover(X, O3) /\ finite(O3)
-          |-
-          ∃(O2, subset(O2, O) /\ cover(Y, O2) /\ finite(O2))
-      ) by RightExists
-
-      // Concluding
-      thenHave(∃(O3, subset(O3, preimages(f, X, Y, O)) /\ cover(X, O3) /\ finite(O3)) |- ∃(O2, subset(O2, O) /\ cover(Y, O2) /\ finite(O2))) by LeftExists
-      have(thesis) by Tautology.from(lastStep, existsO3)
-    }
-    thenHave(openCover(Y, T2, O) ==> ∃(O2, subset(O2, O) /\ cover(Y, O2) /\ finite(O2))) by Tautology
-    val yIsCompact = thenHave(forall(O, openCover(Y, T2, O) ==> ∃(O2, subset(O2, O) /\ cover(Y, O2) /\ finite(O2)))) by RightForall
-
-    have(thesis) by Tautology.from(yIsCompact, yIsTop, compact.definition of (X := Y, T := T2))
   }
 }
